@@ -1,3 +1,11 @@
 module github.com/ldeng7/go-x
 
-go 1.15
+go 1.18
+
+require github.com/smartystreets/goconvey v1.7.2
+
+require (
+	github.com/gopherjs/gopherjs v0.0.0-20181017120253-0766667cb4d1 // indirect
+	github.com/jtolds/gls v4.20.0+incompatible // indirect
+	github.com/smartystreets/assertions v1.2.0 // indirect
+)
